requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pyserial==3.5